==> scripts/02_eda.Rmd <==
---
title: "Exploratory Data Analysis"
output: html_notebook
---

This notebook is just an exploratory data analysis to begin with.

```{r setup}
library(readr)
library(dplyr)
library(ggplot2)
library(DataExplorer)

options(scipen = 999)

train <- read_csv(file.path(here::here(), "data/external/train.csv"))
```

Lets take a look at our data.

```{r a first look}
train %>% head
```

Lets plot the distribution of house prices.

```{r house price distribution}
train %>%
    ggplot(aes(x = SalePrice)) +
    geom_histogram()
```

We will be assessed based upon the RMSE between the logarithm of the predicted value and the logarithm of the observed sales price. Therefore, we probably want to be building a model on the log score.

Lets see the plot above of house price distribution on the log scale.

```{r house price distribution log scale}
train %>%
    ggplot(aes(x = SalePrice)) +
    geom_histogram() +
    scale_x_log10()
```

How many houses are there?

```{r number of houses}
nrow(train)
```

Based upon my knowledge on what influences the price of a house, I would say the main features (and therefore the ones I should spend the most time using in my model) are:

These are:

* Number of bedrooms/rooms
* Amount of land
* Location (urban/rural)
* House or flat
* Condition of the house
* House Age

There are so many variables in our dataset, and I am sure that a lot of them don't give us much information at all. By referring to the data_description.txt file and our data explorer report, lets make a list of the variables to keep.

* `MSZoning` -Tells us the rough location of the property (will be correlated with `Neighbourhood`)
* `LotArea` - Tells us the amount of the land the property has
* `Neighbourhood` - Tells us the location of the neighbourhood of the property
* `BldgType` - Tells us whether it is detached, terraced, apartement or end terrace (not sure what a two-family conversion is)
* `OverallQual` and `OverallCond` - Overall quality and conditon of the property. Should be easy to use in the model as this is just a number between 1 and 10 which is ordered.
* `YearBuilt` and `YearRemodAdd` - Newer properties are generally worth slightly more I think. We probably will only want to use one of these variables as otherwise it would be confusing which to use, however I'm not sure which variable has more importance.
* `TotalBsmtSF`, `1stFlrSF` and `2ndFlrSF` - this tells us how large the interior of the property is
* `MoSold` and `YrSold` - there may be a time series effect of house prices

These variables should allow us to capture all the features that I said above think will have the most effect on the house price.

Once we have created the model some of the other categoric variables such as the roof material, exterior covering, foundation etc may be useful to add in as random effects (only useful though if the categories are fairly evenly distributed).

So in summary, for each of the features that I am trying to estimate, these are the variables that will allow me to create it.

* Number of bedrooms/rooms - sum of TotalBsmtSF, 1stFlrSF and 2ndFlrSF
* Amount of land - LotArea
* Location (urban/rural) - both Neighborhood and MSZoning as random effects
* House or flat - BldgType - can just add this in as another random effect
* Condition of the house - OverallCond, OverallQual may also be useful
* House Age - I don't think this is relevant given the feature above
